refactor(oidc): extract OIDC routes to separate module
Move OIDC-specific code from tokensecurity.js to new oidc-auth.ts module,
addressing code review feedback from @tkurki.

Changes:
- Create src/oidc/oidc-auth.ts with registerOIDCRoutes() and findOrCreateOIDCUser()
- Define OIDCAuthDependencies interface for clear contract between modules
- Add session cookie helpers to tokensecurity.js (setSessionCookie, clearSessionCookie, generateJWT)
- Fix security bug: OIDC callback now includes sameSite and secure cookie options
- Add docs/security-architecture.md documenting how security components fit together

This reduces tokensecurity.js by ~200 lines and improves separation of concerns.
